backend/requirements.txt
fastapi==0.108.0
sqlmodel==0.0.16
uvicorn==0.24.0
psycopg2-binary
python-multipart==0.0.6
pydantic==1.10.18
pydantic-settings
python-jose
passlib[bcrypt]==1.7.4
jinja2==3.1.2
python-dotenv==1.0.0